refactor: vue.config - fix menu-light background-color closed #343
